tests/wpt/web-platform-tests/trusted-types/TrustedHTML.tentative.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="./support/helper.js"></script>
<script>
  test(t => {
    var html = TrustedHTML.escape(STRINGS.unescapedHTML);

    assert_equals(html.toString(), STRINGS.escapedHTML);
  }, "Basic escaping.");

  test(t => {
    var html = TrustedHTML.unsafelyCreate(STRINGS.unescapedHTML);

    assert_equals(html.toString(), STRINGS.unescapedHTML);
  }, "Basic unsafe construction.");
</script>
